+++ scripts/imitation_learning/locomanipulation_sdg/gr00t/data_config.py
@@ -22,7 +22,6 @@ from gr00t.configs.data.embodiment_configs import register_modality_config
 from gr00t.data.embodiment_tags import EmbodimentTag
 from gr00t.data.types import ModalityConfig
 
-
 g1_locomanipulation_sdg_config = {
     "video": ModalityConfig(
         delta_indices=[0],